Organize, Split, Parametrize timezones/timestamps tests (#19473)
